First re-commit
New 'declassed' style.
About to implement 'tranisitions'
Basic effects are flavorFill, flavorWipe, rainbow, rainbowCycle, &
sparkle.
At some point the grid and choreography of the effects will be moved
out to a file that's specific for each project while the core effects
and logic will remain as a library to be included. We have a ways to go.